requirements.txt
pytest
pyyaml
jsonschema[format]
pre-commit
black
pyspellchecker
